move waypoint class to common
