+++ include/scrimmage/common/Waypoint.h
@@ -30,8 +30,8 @@
  *
  */
 
-#ifndef INCLUDE_SCRIMMAGE_PLUGINS_AUTONOMY_WAYPOINTGENERATOR_WAYPOINT_H_
-#define INCLUDE_SCRIMMAGE_PLUGINS_AUTONOMY_WAYPOINTGENERATOR_WAYPOINT_H_
+#ifndef INCLUDE_SCRIMMAGE_COMMON_WAYPOINT_H_
+#define INCLUDE_SCRIMMAGE_COMMON_WAYPOINT_H_
 
 #include <scrimmage/math/Quaternion.h>
 
@@ -81,4 +81,4 @@ class Waypoint {
 };
 } // namespace autonomy
 } // namespace scrimmage
-#endif // INCLUDE_SCRIMMAGE_PLUGINS_AUTONOMY_WAYPOINTGENERATOR_WAYPOINT_H_
+#endif // INCLUDE_SCRIMMAGE_COMMON_WAYPOINT_H_
